+++ server/auth/index.ts
@@ -1,8 +1,10 @@
-import type { User } from "../database/schemas.js";
+import type { MFAOption, User } from "../database/schemas.js";
 import { addJwtToBlacklist, jwtTokenFromRequest, newAccessToken } from "./jwt.js";
 import { asyncWrapper } from "../asyncWrapper.js";
-import { compare, genSalt, hash } from "bcrypt";
+import { compare } from "bcrypt";
 import { Context } from "./Context.js";
+import { generateHash, generateSalt } from "./generators.js";
+import { generateTOTPSecret, verifyTOTP } from "./totp.js";
 import { MAX_USERS } from "./limits.js";
 import { metadataFromRequest } from "./requireAuth.js";
 import { respondSuccess } from "../responses.js";
@@ -11,6 +13,7 @@ import { throttle } from "./throttle.js";
 import { v4 as uuid } from "uuid";
 import {
 	BadRequestError,
+	ConflictError,
 	DuplicateAccountError,
 	NotEnoughRoomError,
 	UnauthorizedError,
@@ -30,14 +33,7 @@ interface ReqBody {
 	newaccount?: unknown;
 	password?: unknown;
 	newpassword?: unknown;
-}
-
-async function generateSalt(): Promise<string> {
-	return await genSalt(15);
-}
-
-async function generateHash(input: string, salt: string): Promise<string> {
-	return await hash(input, salt);
+	token?: unknown;
 }
 
 async function userWithAccountId(accountId: string): Promise<User | null> {
@@ -50,7 +46,7 @@ async function userWithAccountId(accountId: string): Promise<User | null> {
  * Returns a fresh document ID that is virtually guaranteed
  * not to have been used before.
  */
-function newDocumentId(this: void): string {
+function newDocumentId(): string {
 	return uuid().replace(/-/gu, ""); // remove hyphens
 }
 
@@ -61,7 +57,7 @@ function newDocumentId(this: void): string {
  *
  * @see https://thecodebarbarian.com/oauth-with-node-js-and-express.html
  */
-export function auth(this: void): Router {
+export function auth(): Router {
 	return Router()
 		.post<unknown, unknown, ReqBody>(
 			"/join",
@@ -98,7 +94,7 @@ export function auth(this: void): Router {
 				await upsertUser(user);
 
 				// ** Generate an auth token and send it along
-				const access_token = await newAccessToken(req, user);
+				const access_token = await newAccessToken(req, user, []);
 				const { totalSpace, usedSpace } = await statsForUser(user.uid);
 				respondSuccess(res, { access_token, uid, totalSpace, usedSpace });
 			})
@@ -121,22 +117,164 @@ export function auth(this: void): Router {
 				}
 
 				// ** Get credentials
-				const storedUser = await userWithAccountId(givenAccountId);
-				if (!storedUser) {
+				const user = await userWithAccountId(givenAccountId);
+				if (!user) {
 					console.debug(`Found no user under account ${JSON.stringify(givenAccountId)}`);
 					throw new UnauthorizedError("wrong-credentials");
 				}
 
 				// ** Verify credentials
-				const isPasswordGood = await compare(givenPassword, storedUser.passwordHash);
+				const isPasswordGood = await compare(givenPassword, user.passwordHash);
 				if (!isPasswordGood) {
 					console.debug(`The given password doesn't match what's stored`);
 					throw new UnauthorizedError("wrong-credentials");
 				}
 
+				// ** If the user's account has a TOTP secret set and locked-in, validate="totp"
+				const validate: MFAOption | "none" =
+					(user.totpSecret ?? "") && // has a secret
+					user.requiredAddtlAuth?.includes("totp") === true // totp enabled
+						? "totp"
+						: "none";
+
 				// ** Generate an auth token and send it along
-				const access_token = await newAccessToken(req, storedUser);
-				const uid = storedUser.uid;
+				const access_token = await newAccessToken(req, user, []);
+				const uid = user.uid;
+				const { totalSpace, usedSpace } = await statsForUser(uid);
+				respondSuccess(res, { access_token, validate, uid, totalSpace, usedSpace });
+			})
+		)
+		.get(
+			"/totp/secret",
+			asyncWrapper(async (req, res) => {
+				// ** TOTP Registration
+
+				const { user, validatedWithMfa } = await metadataFromRequest(req);
+				const uid = user.uid;
+				const accountId = user.currentAccountId;
+
+				if (
+					validatedWithMfa.includes("totp") ||
+					user.requiredAddtlAuth?.includes("totp") === true
+				) {
+					// We definitely have a secret, the user used it to get here!
+					// Or the user already has TOTP enabled.
+					// Either way, we should not regenerate the token. Throw a 409:
+					throw new ConflictError("totp-conflict", "You already have TOTP authentication enabled");
+				}
+
+				// Generate and store the new secret
+				const secret = await generateTOTPSecret(accountId);
+
+				// We should not lock in the secret until the user hits /totp/validate with that secret.
+				// Just set the secret, not the 2fa requirement
+				await upsertUser({
+					currentAccountId: accountId,
+					passwordHash: user.passwordHash,
+					passwordSalt: user.passwordSalt,
+					requiredAddtlAuth: [], // TODO: Leave other 2FA alone
+					totpSecret: secret,
+					uid,
+				});
+
+				respondSuccess(res, { secret });
+			})
+		)
+		.delete(
+			"/totp/secret",
+			asyncWrapper<unknown, unknown, ReqBody>(async (req, res) => {
+				// ** TOTP Un-registration
+
+				const { user /* validatedWithMfa */ } = await metadataFromRequest(req);
+				const uid = user.uid;
+				const accountId = user.currentAccountId;
+				const secret = user.totpSecret;
+
+				const givenPassword = req.body.password;
+				const token = req.body.token;
+				if (
+					typeof givenPassword !== "string" ||
+					typeof token !== "string" ||
+					givenPassword === "" ||
+					token === ""
+				) {
+					throw new BadRequestError("Improper parameter types");
+				}
+
+				// If the user has no secret, treat the secret as deleted and return 200
+				if (secret === null || secret === undefined || !secret) {
+					respondSuccess(res);
+					return;
+				}
+
+				// Validate the user's passphrase
+				const isPasswordGood = await compare(givenPassword, user.passwordHash);
+				if (!isPasswordGood) {
+					throw new UnauthorizedError("wrong-credentials");
+				}
+
+				// Re-validate TOTP
+				const isCodeGood = verifyTOTP(token, secret);
+				if (isCodeGood) {
+					respondSuccess(res);
+				} else {
+					throw new UnauthorizedError("wrong-credentials");
+				}
+
+				// Delete the secret and disable 2FA
+				await upsertUser({
+					currentAccountId: accountId,
+					passwordHash: user.passwordHash,
+					passwordSalt: user.passwordSalt,
+					requiredAddtlAuth: [], // TODO: Leave other 2FA alone
+					totpSecret: null,
+					uid,
+				});
+
+				// TODO: Re-issue an auth token with updated validatedWithMfa information
+				respondSuccess(res);
+			})
+		)
+		.post<unknown, unknown, ReqBody>(
+			"/totp/validate",
+			asyncWrapper(async (req, res) => {
+				// ** Check that the given TOTP is valid for the user. If valid, but the user hasn't yet enabled a 2FA requirement, enable it
+
+				// Get credentials
+				const { user } = await metadataFromRequest(req);
+				const uid = user.uid;
+				const secret = user.totpSecret;
+
+				const token = req.body.token;
+				if (typeof token !== "string" || token === "") {
+					throw new BadRequestError("Improper parameter types");
+				}
+
+				// If the user doesn't have a secret stored, return 409
+				if (secret === null || secret === undefined || !secret) {
+					throw new ConflictError(
+						"totp-secret-missing",
+						"You do not have a TOTP secret to validate against"
+					);
+				}
+
+				// Check the TOTP is valid
+				const isValid = verifyTOTP(token, secret);
+				if (!isValid) throw new UnauthorizedError("wrong-credentials");
+
+				// If there's a pending secret for the user, and the user hasn't enabled a requirement, enable it
+				if (!user.requiredAddtlAuth || !user.requiredAddtlAuth.includes("totp")) {
+					await upsertUser({
+						currentAccountId: user.currentAccountId,
+						passwordHash: user.passwordHash,
+						passwordSalt: user.passwordSalt,
+						requiredAddtlAuth: ["totp"], // TODO: Leave other 2FA alone
+						totpSecret: secret,
+						uid,
+					});
+				}
+
+				const access_token = await newAccessToken(req, user, ["totp"]);
 				const { totalSpace, usedSpace } = await statsForUser(uid);
 				respondSuccess(res, { access_token, uid, totalSpace, usedSpace });
 			})
@@ -150,7 +288,7 @@ export function auth(this: void): Router {
 
 				const metadata = await metadataFromRequest(req); // throws if bad
 
-				const access_token = await newAccessToken(req, metadata.user);
+				const access_token = await newAccessToken(req, metadata.user, metadata.validatedWithMfa);
 				const uid = metadata.user.uid;
 				const account = metadata.user.currentAccountId;
 				const { totalSpace, usedSpace } = await statsForUser(uid);